@@ -4,9 +4,10 @@
   "description": "PNG plugin for calipers.",
   "main": "lib/index.js",
   "scripts": {
-    "test": "NODE_ENV=test istanbul cover _mocha -- test --recursive --timeout 15000",
-    "enforce": "istanbul check-coverage --statement 100 --branch 100 --function 100 --lines 100",
-    "lint": "eslint ."
+    "test": "NODE_ENV=test nyc mocha -- test --recursive --timeout 15000",
+    "enforce": "nyc check-coverage --statement 100 --branch 100 --function 100 --lines 100",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix ."
   },
   "repository": {
     "type": "git",
@@ -25,15 +26,11 @@
   "homepage": "https://github.com/calipersjs/calipers-png",
   "devDependencies": {
     "chai": "2.x.x",
-    "coveralls": "2.x.x",
-    "eslint": "1.x.x",
-    "eslint-config-lob": "1.x.x",
-    "istanbul": "0.x.x",
-    "mocha": "2.x.x",
+    "eslint": "7.11.0",
+    "eslint-config-lob": "5.2.0",
+    "nyc": "15.1.0",
+    "mocha": "10.0.0",
     "mocha-lcov-reporter": "0.x.x"
-  },
-  "dependencies": {
-    "bluebird": "3.x.x"
   },
   "peerDependencies": {
     "calipers": "2.x.x"
