@@ -1,17 +1,19 @@
 'use strict';
 
-var expect = require('chai').expect;
-var utils  = require('../lib/utils');
+const expect = require('chai').expect;
+const utils  = require('../lib/utils');
 
-describe('utils', function () {
+describe('utils', () => {
 
-  describe('ascii', function () {
-    it('should return the correct string', function () {
-      var string = 'Test String%!@#\n';
-      var buffer = new Buffer(string);
-      var result = utils.ascii(buffer, 0, string.length);
+  describe('ascii', () => {
+
+    it('should return the correct string', () => {
+      const string = 'Test String%!@#\n';
+      const buffer = Buffer.from(string);
+      const result = utils.ascii(buffer, 0, string.length);
       return expect(result).to.eql(string);
     });
+
   });
 
 });
